tdf#161501 icon choice ctrl: Use tab text colors
Instead of using the generic window or highlight
text color, use the specific text colors for tabs
when painting entries of the vertical tab control,
similar to how it's done for the non-vertical
tab control (see `TabControl::ImplDrawItem`).

This is also in preparation for changes to make more
use of the system colors for the entries.

It e.g. ensures that black text color is used for the
text of the focused entry with Breeze (light) in my Debian
testing KDE Plasma 5 setup with Rafael's pending change [1],
preventing the issue of too low contrast mentioned in [2]
that is otherwise seen in my setup, but not Rafael's for some
reason.

[1] https://gerrit.libreoffice.org/c/core/+/170554
[2] https://gerrit.libreoffice.org/c/core/+/170554/comments/f9d0f988_6a0f5eec

Change-Id: I718d08f717bd896b7fca9fc39eedae68fba4bd25
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170677
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins