don't use GetItemSurrogates in SwFrame::GetVirtPageNum
So lets implement this just by walking the document node structure.
I have tried to limit how much of the document this method now scans,
to avoid it also turning into an O(n^2) problem, but I might
have missed something, so
(a) the loop might be too conservative, in which case
performance could be better.
(b) the loop might not go far enough, in which case we might see
bugs with virtual page numbers.

This shaves 10% off the rendering time of large complex docx
documents.

Change-Id: I69711c65197caf278bd3ad2931c4817e72b28c5e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170451
Tested-by: Jenkins
Reviewed-by: Noel Grandin <noel.grandin@collabora.co.uk>