Resolves: tdf#161662 use some hackery to bring bullet combobox height down
There are some Arabic, etc numbering options in there that cause gtk to
allocate a large odd-looking height for the combobox.

What we can try is to remove top/bottom padding for this widget, but
then sync it with a sizegroup to another combobox and then it gets a
more natural looking height.

Change-Id: Idecbb3ee7b61fe55805d26cabdbaf1885574b3cf
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170751
Tested-by: Jenkins
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>