CppunitTest_sw_uiwriter3: avoid font fallback in testTdf133967
I'm not sure when this started, but the testcase fails with:

- Expected: 6
- Actual  : 7

I.e. the initial page count is meant to be more than 1, but we
explicitly assert that page count.

Once gb_CppunitTest_set_non_application_font_use is set to abort for
this suite, then vcl says the font fallback is:

vcl/unx/generic/fontmanager/fontconfig.cxx:1195: PrintFontManager::Substitute: missing font: 'Albany' try: Liberation Sans instead

Do that replacement, that restores the 6 pages on load even in cases
where it was 7 previously.

Change-Id: I0cf40c0cdbbb5c1e894468371ea260f0d15a3e40
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170759
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>