tdf#161741 tdf#161705 pgnum wizard: avoid undo crash (sledgehammer)
This was fixed differently (hopefully more elegantly) in master,
but that patch cannot be confidently backported.
So this alternate approach is proposed for backport to stable.

The "kill undo completely" approach
that this patch is reverting back too is still commonly used for
anything-to-do-with-page-style-headers-and-footers.

i.e sw/source/core/unocore/unostyle.cxx // Fix i64460: as long as
  Undo of page styles with header/footer causes trouble...

LO 7.6 introduced the page number wizard which has more prominently
exposed existing ways to make undo/redo of page style changes
to crash. So treat this as pre-LO 7.2
and simply erase the undo stack after running the wizard.
------------------------------------------------------------------
The ability to undo/redo the "Header/footer change" started in
bibisect-linux-64-7.2 commit 910d9a081daf11ea53235d227c3dc9623942a824
commit 8d8486f43c1a8a51157bfc3e0b87090b05a9229e
Author: Daniel Arato (NISZ) on Mon Feb 22 16:59:38 2021 +0100
    tdf#46561 sw: fix lost undo stack setting header/footer

-   if( bHeaderFooterChanged )
-       GetIDocumentUndoRedo().DelAllUndoObj();

Change-Id: I00c529784bfd7b2cd1230474c05eea8516e6c1e3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172822
Reviewed-by: Justin Luth <jluth@mail.com>
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/172911