Related: tdf#162072 DOCX import: handle font family for characters
Open the bugdoc, the first paragraph is meant to be sans and the second
paragraph is meant to be serif, but both of them are serif.

The bugdoc uses the fonts 'IBM Plex Sans' and 'IBM Plex Serif', which is
not something we bundle, so the font fallback is expected, but the
fallback should have a matching font family (roman vs swiss).

Fix the problem by implementing support for <w:family w:val="..."> in
the DOCX import, which was just ignored previously.

Now the DOCX import result in on par with the ODT import result.

Change-Id: I321b9fc6f63126ca5ad61af57af8a5bc7456d5b4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170772
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>