tdf#161741 tdf#161705 undo: delay ClearRedo until Undo finishes
I know NOTHING about the intricacies of Undo/Redo.
However, this is my feeble attempt to add some sanity to it.
It should not be the responsibility of the caller
to know when they are allowed to call ClearRedo.

This patch reverts commit 0cd000bb83719982c1fd2265ea040c82af5bf98e
author  Daniel Arato (NISZ) on Mon May 10 15:41:28 2021 +0200
    tdf#141613 sw: avoid possible crash when undoing header creation
which was not a sufficient hack.

I hope this patch lays a better framework
to handle future similar issues.

vvv NAIVE OPTIMIST ALERT vvv

mbDoing was added with
commit 9cb53122e9e098bc8a6bf53c14b18415e369dd6d
Author: Frank Schoenheit on Fri Oct 22 15:00:39 2010 +0200
    undoapi: more I/SfxUndoManager changes
    in preparation of new Undo API features

and has been untouched since then AFAICS,
and there was never any mechanism to change mbDoing.
In other words, it has been sitting there doing NOTHING.

So, I am taking it over and using it how I imagine it was intended,
and how it is documented.

Change-Id: I7aa355cc6458ac8ba34ddb9ee73fc850dcbca702
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170793
Tested-by: Jenkins
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>