Resolves tdf#162018 Unexpected loss in focus in the "Manage Changes"
window when adding a change in the document (Track changes mode),
round 3

Done to address loss of selection when an entry is added to the
beginning of the changes list. The patch reworks round 1 and 2 to
handle any entry insert position. It also keeps the selected entry
selected after a sort by document order.

Change-Id: Icb6f6b6b80ad9ddc5d008a00d195ceb7c1d6dc3f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170828
Tested-by: Jenkins
Reviewed-by: Jim Raykowski <raykowj@gmail.com>