tdf#161501 vcl: Make TabitemValue position-aware
Add a new `TabBarPosition` enum that describes whether
a tab bar is at the top, left, right, or bottom
(relative to the tab page content) and make use of
that for the new `meTabBarPosition` member for
`TabitemValue` that gets used for drawing tab items
via the native rendering API.

Set the position accordingly for both the vertical
tab control (`TabControl`) and `SvxIconChoiceCtrl_Impl`,
which implements the tab bar for `VerticalTabControl`.

For Qt-based VCL plugins, implement mapping to
the corresponding `QTabBar::Shape`. [1]

At least the Breeze Qt style uses that shape to decide
whether on what side of the tab item to draw the
selection/focus indicator.

This change by itself doesn't yet make any visual
difference for the Breeze style, as the "Top" value/
`QTabBar::RoundedNorth` is the default anyway, and
`SvxIconChoiceCtrl_Impl` doesn't yet use the
native drawing API for drawing focus/selection, but
only for mousehover feedback.
The plan is to implement that in a follow-up change.

[1] https://doc.qt.io/qt-6/qtabbar.html#Shape-enum

Change-Id: Idf1e41d3bc309d152a4a36d14e8617bd6429940c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170841
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>