tdf#161501 icon choice ctrl: Drop extra drawing of entry bg
Drop extra drawing of a rectangle with the background
color just for selected entries.

It's unclear why that is done, and why it's done
only for selected entries, since the background
color used here is unrelated to actual selection
(e.g. is just white when using the gen VCL plugin
or kf5 with the light Breeze theme).

Drawing of the actual selection background
happens further down.
Therefore, drop drawing that extra rectangle
to simplify code and having one place less
to care about when switching to using native
drawing for the whole entry in an upcoming
commit.

I see no visual difference with or without
this commit in place with either gen or kf5
with either the "Insert -> "Hyperlink" or the
"Format" -> "Page Style" vertical tab pages
dialogs in Writer.

Change-Id: Ia1a7275238a919d245420b5ab4abaf7b3e85bcf0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170843
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>