tdf#161501 icon choice ctrl: Draw whole entry natively
When native drawing of tab items is supported,
natively draw the whole (vertical) tab item (except for the
text) in `SvxIconChoiceCtrl_Impl::PaintEntry`,
instead of using native drawing API only for highlighting
the mouse-hovered entry, as originally introduced in

    commit 8708e8cf90d0ac0acbbce34b82f10ef7352d5062
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Jul 11 10:47:50 2024 +0200

        tdf#161026 tdf#161355 tdf#161501 icon choice ctrl: Natively draw mouse-hover

        ... feedback.

Set the corresponding state flags for selected and
focused entry, so that the native drawing routine
can take care of drawing the selected/focused
entry accordingly instead of using the native drawing
API to draw a `ControlType::WindowBackground` for that
purpose.

With this commit in place, all entries (not just the
mouse-hovered, focused or selected) entry are now
drawn using the native drawing API.

For the kf5 VCL plugin with the Breeze style, this
makes them look similar to the entries for the
non-vertical tabbar used elsewhere in LibreOffice,
or actual native `QTabBar`s.

Due to

    Change-Id: Idf1e41d3bc309d152a4a36d14e8617bd6429940c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jul 22 08:25:32 2024 +0200

        tdf#161501 vcl: Make TabitemValue position-aware

, the selection indicator is drawn vertically as expected
when using the Breeze style.

Restrict manually drawing entries (including focus
and selection indicator) to the case where native drawing
is not implemented (e.g. the gen VCL plugin).

Further tweaking the visual appearance (e.g.
getting rid of extra white margins and space between the
entries for the variant where icons are used,
like the "Insert" -> "Hyperlink" dialog in Writer) can be
done in separate commits.

Change-Id: Ie0c8ba1b56f6bb76ece6761253b93a109bb3a3f4
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170847
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins