tdf#161501 icon choice ctrl: Drop extra space between items
Drop the extra spacing between vertical tab bar items, which
became quite noticeable as extra white area between the otherwise
light-grey entries when using the Breeze (light) style and the
kf5 VCL plugin for the "Insert" -> "Hyperlink" dialog now that
the tab items are drawn natively since

    Change-Id: Ie0c8ba1b56f6bb76ece6761253b93a109bb3a3f4
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jul 22 11:05:30 2024 +0200

        tdf#161501 icon choice ctrl: Draw whole entry natively

and which is also more visible for non-natively drawn items
since

    Change-Id: I578755a8a82ea811765b2e0dc3815fb47f67863c
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Tue Jul 23 10:41:16 2024 +0200

        tdf#161501 icon choice ctrl: Draw selection within item bounds

With this in place, the extra margin between entries
in tab items with icons in the "Insert" -> "Hyperlink"
dialog in Writer is gone for both, the gen VCL plugin
and the kf5 one (tested with the Breeze style).

For the gen VCL plugin, there's still a small gap between
text-only items (e.g. in the "Format" -> "Page Style" dialog
which has a different cause and will be addressed in a separate
commit.

Change-Id: Iac34098ef8d0f90f1ac0844df63839a6a99b83a7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170850
Reviewed-by: Rafael Lima <rafael.palma.lima@gmail.com>
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>