tdf#161501 icon choice ctrl: Draw selection within item bounds
In the same way that the calculated focus rect of an item
is used for drawing selection and focus indicators, use
it for mouse-hover highlighting as well, instead of
drawing a rectangle that reaches beyond the bounds
of the item itself for the non-native drawing path
of vertical tab bar items.

With this in place, hovering over an entry other than
the selected/focused one now shows a rectangle of the same
size as that for the focused entry, which is more
consistent and also aligns this with the case
where the native drawing API for drawing the item is used,
e.g. on Windows or with the kf5 VCL plugin on Linux.

This also means that a small gap can now be seen
between selected item and mouse-hovered item when
hovering over an item next to the currently selected
one for the gen VCL plugin as well, just as was
already the case before for the kf5 VCL plugin.
(Dropping that extra margin will happen in a
separate commit.)

Change-Id: I578755a8a82ea811765b2e0dc3815fb47f67863c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170891
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>