tdf#161501 icon choice ctrl: Drop debug-only mode switch feature
Drop code that allows switching the text mode of
the currently selected tab item in a vertical tab
bar using Shift+F10 or Ctrl+F10, which is only
available in dbgutil builds, and was therefore
likely meant to be used for debugging purposes
only.

As related code will be changed in a follow-up
commit, this would have to be adjusted to keep
it working. Instead of spending time on that,
just drop it. Testing how using another mode is still
possible e.g. by locally changing the value assigned
to `SvxIconChoiceCtrlEntry::eTextMode` in the
`SvxIconChoiceCtrlEntry` ctor instead.

Change-Id: If309f84fcd9a99337bc5896ce5c3238333427da5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170892
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>