tdf#161501 icon choice ctrl: Don't modify bound rect to center content
Don't modify the bound rect of tab items used used e.g.
for drawing the focus/selection/mouse-hover rectangles/indicators
in order to center the content horizontally and vertically.

Making the bound rect smaller in vertical direction
in `SvxIconChoiceCtrl_Impl::Center` resulted in margins
between entries for the non-native drawing path (e.g. the
gen VCL plugin on Linux) for the text-only case
(e.g. "Format" -> "Page Style" dialog in Writer).

Making sure that the content is centered should rather
happen in the drawing routine.

Horizontally centering already happens for the icon-case
anyway, as the `PAINTFLAG_VER_CENTERED` gets passed as flag
to the `PaintItem` call for the `WB_ICON` case in
`SvxIconChoiceCtrl_Impl::PaintEntry` anyway.

And for text-only mode, which was apparently the only one
for which centering vertically in `SvxIconChoiceCtrl_Impl::Center`
had an effect, vertically drawing text was implemented in the drawing
code path in

    commit 180f0c1ec8e195043b5f4298737a219026a8b944
    Author: Rafael Lima <rafael.palma.lima@gmail.com>
    Date:   Tue Jul 16 20:52:28 2024 +0200

        tdf#161501 Improve visuals of selected entries in Vertical tabs

. This commit addresses the

>   For the gen VCL plugin, there's still a small gap between
>   text-only items (e.g. in the "Format" -> "Page Style" dialog
>   which has a different cause and will be addressed in a separate
>   commit.

aspect mentioned in earlier commit

    Change-Id: Iac34098ef8d0f90f1ac0844df63839a6a99b83a7
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Mon Jul 22 11:39:48 2024 +0200

        tdf#161501 icon choice ctrl: Drop extra space between items

Change-Id: I2077efc8ce1e97d52cf078d63bb4e4b528e389eb
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170895
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins