tdf#162010 match based on key equivalent instead of key event
The original fix for tdf#49853 only looked worked in the
case when both a key shortcut was pressed and the resulting
key event was not an input method event. If either of these
conditions weren't true, tdf#49853 would still occur.

Since we know which menu item is being triggered, check if
this menu item's key equivalent has been set to one of the
edit actions.

This change basically expands the fix for tdf#49853 to
include all cases that trigger a menu item, not just simple
key events.

Also, the fix for tdf#126638 would also fail when a key
shortcut was pressed in a modal window if the resulting key
event was an input method event.

Change-Id: Iaa2e06affe0e347380a383755684568fcdc26b9a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170908
Tested-by: Jenkins
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>