tdf#161501 icon choice ctrl: Actually set text colors for non-selection
    commit 24103bdf3fa3d6962efcc6fb6ee5c95b62e29cc4
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Thu Jul 18 08:58:10 2024 +0200

        tdf#161501 icon choice ctrl: Use tab text colors

        Instead of using the generic window or highlight
        text color, use the specific text colors for tabs
        when painting entries of the vertical tab control,
        similar to how it's done for the non-vertical
        tab control (see `TabControl::ImplDrawItem`).

        (...)

had added code to set the font color for the different
states of a vertical tab control item, but as this
was in the `if (bSelected)` block, only the one
for selection was actually used, so the other cases
were consequently removed in

    commit a4befb29f325d4800b3408abb6e6df608a97ecf9
    Author: Caolán McNamara <caolan.mcnamara@collabora.com>
    Date:   Sat Jul 20 19:20:13 2024 +0100

        cid#1610738 Logically dead code

again.

Thanks to Chris Sherlock for pointing this out!

Reintroduce setting colors, but drop the `bSelected`
condition altogether. Drop explicitly setting a fill
color, which the implementation for the non-vertical
tab control in `TabControl::ImplDrawItem` also doesn't
do and which would e.g. result in an odd grey background
for non-selected entries when used unconditionally with
the gen VCL plugin.

I see no visual difference with or without this commit
in place on Windows or for the gen VCL plugin or the kf5 VCL
plugin with the Breeze style on Linux, but other theming/styles
could depend on according colors being set.

Change-Id: Iccf7170f2de04cead23607977ac8cf7acbc471f3
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170926
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins