fix(invert): Use separate view IDs when inverted
Previously in Online the document background inversion was tied together
such that if someone inverted the background, everyone in dark theme
would end up with an inverted background.

By adding whether the document background is dark to the view render
state, we can create a new view when the background is inverted, and
avoid everyone getting an inverted view.

Change-Id: I0aed702185e642f631854b1f8234355c9e69ff6e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171045
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>