test: Refactor pixel assertion
When a test fails, the line the assertion failed on is printed.
Unfortunately, in the case of this assertion, it was always inside the
same function.

If we instead change the function to only return the pixel color, and
perform the assertion inside the test, assertion failure error messages
are more helpful

Change-Id: Ia56da1b245882c8ddbf21ad98f69facc36d3ec3a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171593
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins