No LibreOfficeKit headers and examples for macOS
LibreOfficeKit has issues in macOS, as discussed in LibreOffice
development mailing List and also in tdf#145127.

LibreOfficeKit unrunnable on macOS
https://lists.freedesktop.org/archives/libreoffice/2024-July/092256.html

Therefore, let's not ship the LibreOfficeKit headers and examples with
LibreOffice Community for macOS. If the above issue is solved, then it
will be possible to ship those again later.

Change-Id: I5ba62ee84432a09732d96a5d86eb0a1bf866671f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171095
Reviewed-by: Patrick Luby <guibomacdev@gmail.com>
Tested-by: Jenkins
Reviewed-by: Hossein <hossein@libreoffice.org>