tdf#162219: pass correct DecimalDigits value to SQLBindParameter
... when setting DECIMAL / NUMERIC.

Passing 'scale' from setObjectWithInfo to SQLBindParameter was wrong
anyway: in setObjectWithInfo, the scale is needed if it converts the
object 'x' to the requested format before sending to the database.

In this case, it mismatched with the actual count of decimals in the
string sent to ODBC: it could be "1" or "1.2"; but 'scale' is always
equal to the scale of the parameter in the database (4 in the bugdoc),
which resulted in the error returned by driver.

Note that there is code that truncates excessive decimals - it is in
OKeySet::impl_convertValue_throw, implemented for i#106772; it would
be best to remove that code, and let the database do its magic: e.g.,
passing a string like "1.99999" to a DECIMAL(19,4) field in MS SQL
Server would round it to 2, while our code truncates it to 1.9999.
The ODBC driver can handle more digits itself. But that isn't in the
scope here, and may need to tweak other database connectors.

Change-Id: Ib50c6d78bfd0cbf5ecd59f46f300107076ec0037
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171127
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>