tdf#158510 Don't change the initial URL when using fallback
The current aURL conversions in the file "unoshap4.cxx"
result in the URL obtained when importing the file being replaced with a
new one formed from the presentation directory and file name.
As a result, the model changes and saving is requested
even if no changes were made to the presentation.

To fix this problem, we added saving the
presentation directory URL into the context for
transfer from SlideImpl to ViewMediaShape.

Regression after: fe897b36aef28dfe175461c43614e22588fcfd84

Change-Id: Ifda8610d0a4874ffa17e759fd45ddbbd2f934c0c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/170528
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>