cool#9654 fix: Skip invalidation if it won't go through
On Android, we often hit this condition where we should get an
invalidation. Unfortunately, the invalidation is skipped due to being in
LOK Tiled Painting, leaving us to deal with whatever tile was rendered
on this go-around...

...Also because of this condition, the tile rendered here is broken.
pFrame->PaintSwFrame is never called, leading to a tile which renders
without things like its text. Normally, this wouldn't be a problem as
it's about to get invalidated, and is more of a nice performance
optimization...

...However given we don't invalidate it, we really see that tile without
any text. Similar workarounds were considered (such as skipping the
condition in another case that appears to be true on Android or
rendering the tile anyway, even if we're about to invalidate it) but
this seems to be the safest bet to avoid something like this happening
on other platforms...

...The true solution is still elusive. It probably includes figuring out
why this happens on Android without happening on other platforms, and
fixing that, as well as perhaps figuring out what to do about this
condition if we *do* genuinely trigger it on Android - as the previous
solution of invalidating everything nearby won't work

Change-Id: Id58f5bae8ae357d116c5f2345e88ec3364cb2172
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171578
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins