feat(invert): Allow inverted background on init
Previously for Online there was no way to save the background invert
state and reload it. Worse, when someone changed the state it would
become the default for new document loads.

This patch allows Online to specify whether it wants the background to
be inverted, which should allow smooth tab refreshes while also avoiding
mingling state from different people.

There is a change to online to support this here:
https://github.com/CollaboraOnline/online/pull/9652

Change-Id: I8c22c03d3b4589736d48509004f7789dd5166389
Refs: https://github.com/CollaboraOnline/online/pull/9652
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171208
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Caolán McNamara <caolan.mcnamara@collabora.com>