tdf#124341: special-case applying autoformat to whole rows
The problem of allocating thousands of columns can be usually avoided
when whole rows are selected (e.g., whole sheet). ApplyPatternArea
(which does the real format application) has a special processing for
'nEndCol == GetDoc().MaxCol()' case, when it only changes the default
beyond the already allocated columns. So we can try to pass a single
range per row (or per row range) where possible, instead of applying
the format to body without the right column, which would disallow the
mentioned optimization.

To make that possible, this change detects if whole rows are selected,
and then it assumes that the user doesn't need the rightmost column
format of the autoformat. This change also rearranges application of
format to be by rows rather than by columns where possible. This is
expected to not slow down the cases that allow by-column application,
because these cases would either apply to whole body block, or not
allow by-row application (alternating columnar pattern).

Indeed, this won't help when the autoformat has checkerboard pattern.
Also, this won't prevent following slow save of the resulting file,
which would be a separate problem.

Change-Id: Id18a0edb7d6d09fccc3ebb1de3c12eaa45b34501
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171223
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins