Fix OTools::ThrowException
First, it used 5-character buffer for SQL state, where MS documentation
for SQLGetDiagRec specifies, that the buffer is for five-character code
plus terminating NULL. The declaration of the function in Windows Kit's
sqlucode.h uses decoration '_Out_writes_opt_(6)' for the parameter.

Next, SQLGetDiagRec may fail or return SQL_NO_DATA;  and then, the data
in buffers will be undefiled.

Change-Id: Ia36555b4fe4f49c8803a25f8fc1cc1a3b6ac4ddc
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171238
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>