tdf#155447 a11y: Set accessible ID for WeldEditView
Let `WeldEditAccessible` derive from `XAccessibleContext2`
and implement `XAccessibleContext2::getAccessibleId`
by returning the accessible ID set in the underlying
drawing area.

This makes an accessible ID of "sentence" show up
in Accerciser for the corresponding edit view in
Writer's spell check dialog when using the qt6 VCL plugin,
since that is the ID set in the .ui file
(cui/uiconfig/ui/spellingdialog.ui).

It's not set for gtk3 yet, which uses native GTK
widgets, and where the accessible ID is not set for the
drawing area yet.

Change-Id: I5cf8958118c84c01ecb3545381a05b306ca53163
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171320
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins