tdf155447 gtk3 a11y: Report accessible ID for drawing area
Override `GtkInstanceWidget::get_accessible_id()`
for `GtkInstanceDrawingArea`, similar to what
is done for the handling of the accessible name
and description.

While accessible name and description don't
have to explicitly be set on the AtkObject
because they can be retrieved from the
underlying XAccessible on demand as
`AtkObjectClass::get_name` and `AtkObjectClass::get_description`
are overriden (s. LO's `atk_object_wrapper_class_init`, and
the ATK implementations of `atk_object_get_name`
and `atk_object_get_description` [1]),
there's no such function to override for the
accessible ID (s. ATK's `atk_object_get_accessible_id`),
so explicitly set it for the drawing area's
AtkObject each time the accessible is queried in
`GtkInstanceDrawingArea::GetAtkObject`.

With this in place, the accessible ID
"sentence" shows up in Accerciser with gtk3 for
the corresponding edit view in Writer's spell check
dialog, as is already the case for qt6 since

    Change-Id: I5cf8958118c84c01ecb3545381a05b306ca53163
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jul 31 13:25:45 2024 +0200

        tdf#155447 a11y: Set accessible ID for WeldEditView

[1] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/fddfd5ae5d44b09c3e587f41f450e05b95262bd9/atk/atkobject.c#L686-729
[2] https://gitlab.gnome.org/GNOME/at-spi2-core/-/blob/fddfd5ae5d44b09c3e587f41f450e05b95262bd9/atk/atkobject.c#L1717-1733

Change-Id: Ib1e889a06dccf4522148d8a17b62259dc08e2b85
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/171321
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins